move onExploreMenuCollapse logic to onQuerySuccess callback to clarify intent
