+++ cmd/ui/src/views/Explore/ExploreSearch/ExploreSearch.tsx
@@ -28,12 +28,15 @@ import {
     PathfindingSearch,
     cn,
     encodeCypherQuery,
+    isGraphResponse,
     useCypherSearch,
     useExploreParams,
+    useExploreSelectedItem,
     useNodeSearch,
     usePathfindingFilters,
     usePathfindingSearch,
 } from 'bh-shared-ui';
+import { FlatGraphResponse, GraphResponse } from 'js-client-library';
 import React, { useState } from 'react';
 import { setAutoRunQueries, setTimeoutSetting } from 'src/ducks/global/actions';
 import { useAppDispatch, useAppSelector } from 'src/store';
@@ -82,7 +85,8 @@ const ExploreSearch: React.FC = () => {
     /* Hooks */
     const classes = useStyles();
 
-    const { exploreSearchTab, setExploreParams } = useExploreParams();
+    const { cypherSearch, exploreSearchTab, setExploreParams } = useExploreParams();
+    const { clearSelectedItem, setSelectedItem } = useExploreSelectedItem();
 
     const nodeSearchState = useNodeSearch();
     const pathfindingSearchState = usePathfindingSearch();
@@ -174,6 +178,25 @@ const ExploreSearch: React.FC = () => {
         dispatch(setTimeoutSetting(disableTimeout));
     };
 
+    const handleQuerySuccess: (data: GraphResponse | FlatGraphResponse) => void = (data) => {
+        if (isGraphResponse(data)) {
+            const returnedNodes = Object.keys(data.data.nodes || {});
+
+            const keepSearchMenuOpenBecauseNoCypherQuery = !cypherSearch && exploreSearchTab === 'cypher';
+            const shouldCloseMenu = !keepSearchMenuOpenBecauseNoCypherQuery && returnedNodes.length >= 1;
+
+            if (returnedNodes.length > 1) {
+                clearSelectedItem();
+            } else if (returnedNodes.length === 1) {
+                setSelectedItem(returnedNodes[0]);
+            }
+
+            if (shouldCloseMenu) {
+                setShowSearchWidget(false);
+            }
+        }
+    };
+
     return (
         <div data-testid='explore_search-container' className='h-full min-h-0 w-[600px] flex gap-4 flex-col rounded'>
             <div
@@ -231,7 +254,7 @@ const ExploreSearch: React.FC = () => {
                             setAutoRun={handleAutoRunChange}
                             disableQueryLimit={disableTimeout}
                             setDisableQueryLimit={handleDisableTimeoutChange}
-                            onExploreMenuCollapse={() => setShowSearchWidget(false)}
+                            onQuerySuccess={handleQuerySuccess}
                         />,
                         /* eslint-enable react/jsx-key */
                     ]}
